UAST: utilize creation of KDoc simple name reference
