+++ uast/uast-kotlin-base/src/org/jetbrains/uast/kotlin/expressions/KotlinStringUSimpleReferenceExpression.kt
@@ -1,10 +1,8 @@
 package org.jetbrains.uast.kotlin
 
-import com.intellij.openapi.components.ServiceManager
 import com.intellij.psi.PsiElement
 import com.intellij.psi.PsiNamedElement
 import com.intellij.psi.ResolveResult
-import org.jetbrains.kotlin.kdoc.psi.impl.KDocName
 import org.jetbrains.kotlin.psi.KtElement
 import org.jetbrains.uast.UElement
 import org.jetbrains.uast.UMultiResolvable
@@ -30,12 +28,3 @@ class KotlinStringUSimpleReferenceExpression(
 
     override fun multiResolve(): Iterable<ResolveResult> = referenceAnchor?.multiResolveResults().orEmpty().asIterable()
 }
-
-fun createKDocNameSimpleNameReference(parentKDocName: KDocName, givenParent: UElement?): USimpleNameReferenceExpression? =
-    parentKDocName.lastChild?.let { psiIdentifier ->
-        parentKDocName.getQualifiedName().lastOrNull()?.let { qualifierText ->
-            val baseResolveProviderService =
-                ServiceManager.getService(psiIdentifier.project, BaseKotlinUastResolveProviderService::class.java)
-            KotlinStringUSimpleReferenceExpression(qualifierText, givenParent, baseResolveProviderService, psiIdentifier, parentKDocName)
-        }
-    }
